refactored the api requests to the server side so that github token is not compromised
